Two strategies to minimize array changes
This brings two changes to prevent "flicker" on arrays.

All observers that use the "non-replacing-array" decoarator will now use
a shortest path algorithm to apply the minimal sequence of splices to
the target array when the source array is replaced.  This will reduce
flicker for sorting and, to a lesser extent, reversing.

The filter block observer now short-circuits if it sees that the
predicates for a range change are identical.  This greatly reduces
activity on the output array.  Unfortunately, it still doesn't aggregate
changes.  Each filter predicate change will cause an individual range
change on the target array.