Fix regression in filter output
In a filter block observer, we cannot short circuit a range change just
because the predicates did not change.  The predicates can be the same
even when the underlying content *does*, in which case we certainly do
need to propagate a range change to the output.  But we still want
to short circuit if the output would not have been changed by the swap.

This patch fixes the regression while also preventing flickering.