Do not throw errors
It is now clear that FRB should not throw errors if it encounteres
invalid input.  It must propagate null instead.  This is because FRB
inputs do not necessarily change atomically.  The result is that state
must be made consistent by the end of a turn (not enforced), but may
pass through invalid states internally.  As such, throwing an exception
would interfere.